slimmed own project structure
