+++ .gitignore
@@ -4,3 +4,4 @@ composer.lock
 index.php
 notes.php
 todo.txt
+/src/Removed
