--- src/USPS/Facades/Usps.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace ctwillie\Usps\Facades;
-
-use Illuminate\Support\Facades\Facade;
-
-class Usps extends Facade {
-
-    /**
-     * Get the registered name of the component.
-     *
-     * @return string
-     */
-    protected static function getFacadeAccessor() { 
-		
-		return 'usps';
-	
-	}
-
-}
