+++ src/Usps/API.php
@@ -11,36 +11,40 @@ abstract class API {
 
     const PROD_URL = 'https://secure.shippingapis.com/ShippingAPI.dll';
     const LOCAL_URL = 'http://production.shippingapis.com/ShippingAPITest.dll';
-    protected $env = 'local';
-    protected $apiUrl = self::LOCAL_URL;
+    protected $env;
+    protected $apiUrl;
     protected $userId;
     
     public function __construct() {
         $this->setEnvironment();
-        $this->setApiUrl();
+        $this->setUrl();
         $this->userId = config('services.usps.userid');
     }
 
     // Determines the current application environment
     public function setEnvironment()
     {
-        // config settings take precedence
         if (Config::has('services.usps.env')) {
             $this->env = strtolower(config('services.usps.env'));
-        // next check .env
         } elseif (App::environment(['local', 'development'])) {
             $this->env = 'local';
-        } elseif (App::environment(['production'])) {
-            $this->env = 'production';
-        }
-        // defaults to local
+        } 
+        $this->env = 'production';
     }
 
-    public function setApiUrl()
+    public function setUrl()
     {
         $this->apiUrl = $this->env === 'local' ? self::LOCAL_URL : self::PROD_URL;
     }
 
+    public function checkForUserId()
+    {
+        if (is_null($this->userId)) {
+            throw new \Exception('USPS: A user ID is required. None found in config/services.php');
+        }
+    }
+
+    // create a map to determine the rootElementName foreach api type
     protected function convertToXML()
     {
         $xml = ArrayToXml::convert( $this->getRequestData(),
@@ -54,23 +58,25 @@ abstract class API {
         return urlencode($xml);
     }
 
-    public function validate()
-    {
-        return $this->makeRequest();
-    }
-
-    abstract public function getRequestData() : array;
-
     public function makeRequest()
     {
+        $this->checkForUserId();
 
         $xml = $this->convertToXML();
         
-        $client = new Client(['base_uri' => $this->apiUrl, 'verify' => config('services.usps.verifyssl')]);
+        $client = new Client(['base_uri' => $this->apiUrl, 'verify' => config('services.usps.verifyssl', true)]);
 
-        $response =  $client->request('GET', "ShippingAPI.dll?API=$this->apiType&XML=$xml");
+        $response =  $client->request('GET', "?API=$this->apiType&XML=$xml");
 
         return $response->getBody();
+    }
+
+    // Abstract to be defined by each entity
+    abstract public function getRequestData() : array;
+
+    public function validate()
+    {
+        return $this->makeRequest();
     }
 
 }
